refactor manifest functionality
